(MODULES-5405) add heredoc and interp detection
This adds detection for here heredocs and interpolation in strings not already decorated. If it finds an undecorated string with interpolation, it skips over it. If it finds an undecorated string with heredoc, it decorates the heredoc tag simliar to the way they're decorated in ruby. We may also want to check for these things in already decorated strings as a warning to say 'hey, you've decorated this but its not going to work because youve got interpolation in there'.